move time selectors
