+++ src/sections/Reports/index.js
@@ -19,6 +19,7 @@ import {
 
 import { getDurationRounding, getDurationAmount, getRoundingOn } from 'sections/Settings/selectors';
 import { sortedProjects } from 'sections/Projects/selectors';
+import { getAllTimeEntries } from 'sections/TimeSheet/selectors';
 
 import ActionMenu from './components/ActionMenu';
 import ReportCharts from './components/Charts';
@@ -30,7 +31,6 @@ import SaveModal from './components/SavedReports/Save';
 import LoadModal from './components/SavedReports/Load';
 
 import { getById } from './selectors';
-import { getAllTimeEntries } from '../../selectors/time';
 
 function toggleFilter(filters: Array<string | null>, filter: string | null) {
   if (filters.indexOf(filter) > -1) {
