+++ src/sections/TimeSheet/components/DateRange/index.js
@@ -8,7 +8,7 @@ import Menu from 'semantic-ui-react/dist/commonjs/collections/Menu';
 
 import { changeDateRange } from '../../../../actions/time';
 
-import { getDateRange } from '../../../../selectors/time';
+import { getDateRange } from '../../selectors';
 
 type DateRangeProps = {
   vertical?: boolean;
